+++ custom_components/supernotify/delivery.py
@@ -39,6 +39,11 @@ from .const import (
     INCLUSION_DEFAULT,
     INCLUSION_FALLBACK,
     INCLUSION_FALLBACK_ON_ERROR,
+    RESERVED_DELIVERY_NAMES,
+)
+from .hass_api import HomeAssistantAPI
+from .model import ConditionVariables, DeliveryConfig, EntityCategory, SelectionRule, Target
+from .options import (
     OPTION_DATA_KEYS_EXCLUDE_RE,
     OPTION_DATA_KEYS_INCLUDE_RE,
     OPTION_DATA_KEYS_SELECT,
@@ -52,12 +57,9 @@ from .const import (
     OPTION_TARGET_CATEGORIES,
     OPTION_TARGET_INCLUDE_RE,
     OPTION_TARGET_SELECT,
-    RESERVED_DELIVERY_NAMES,
     SELECT_EXCLUDE,
     SELECT_INCLUDE,
 )
-from .hass_api import HomeAssistantAPI
-from .model import ConditionVariables, DeliveryConfig, EntityCategory, SelectionRule, Target
 from .static_config import TRANSPORT_NAMES
 
 if TYPE_CHECKING: