+++ tests/components/supernotify/test_media_grab.py
@@ -24,7 +24,6 @@ from custom_components.supernotify.const import (
     CONF_PTZ_DELAY,
     CONF_PTZ_PRESET_DEFAULT,
     CONF_SNAP_WAIT,
-    MEDIA_OPTION_REPROCESS,
     PTZ_DELAY_DEFAULT,
     PTZ_METHOD_FRIGATE,
     PTZ_METHOD_ONVIF,
@@ -47,6 +46,7 @@ from custom_components.supernotify.media_grab import (
     write_image_from_bitmap,
 )
 from custom_components.supernotify.notification import Notification
+from custom_components.supernotify.options import MEDIA_OPTION_REPROCESS
 from custom_components.supernotify.schema import MEDIA_SCHEMA
 
 from .hass_setup_lib import TestingContext